backend/requirements.txt
Django
djangorestframework
djoser
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
celery
redis
email-validator
dnspython
gunicorn
